Refactored other API calls into their own services
