+++ frontend/src/components/DataView/DataView.tsx
@@ -6,42 +6,8 @@ import { Box, TextField } from "@mui/material";
 import { Funnel, MapPin, MapPinLine } from "@phosphor-icons/react";
 import { MapContext } from "../../contexts/MapContext";
 import LoadDataButton from "./LoadDataButton";
-import axios from "axios";
-import { getAPIGatewayURL } from "../../services/metadataService";
-
-const loadLocationData = async (): Promise<
-  LocationDataResponse | undefined
-> => {
-  const requestBody = {
-    datasetId: "example_dataset",
-    location: [
-      { latitude: 51.509865, longitude: -0.118092 }, // Example coordinate
-    ],
-  };
-
-  try {
-    const response = await axios.put<LocationDataResponse>(
-      getAPIGatewayURL() + "/api/loadLocationData",
-      requestBody
-    );
-    return response.data;
-  } catch (error) {
-    console.error("Error loading location data", error);
-    return undefined;
-  }
-};
-
-interface LocationDataResponse {
-  currentDatasetData: DatasetItem[];
-  generalData: DatasetItem[];
-  extraRows: DatasetItem[];
-}
-
-interface DatasetItem {
-  key: string;
-  value: string;
-  mapId: string;
-}
+import { LocationDataResponse } from "../../types/LocationDataTypes";
+import { fetchLocationData } from "../../services/locationDataService";
 
 function DataView() {
   const { currentTabsCache } = useContext(TabsContext);
@@ -73,14 +39,19 @@ function DataView() {
     }
   }, [currentMapCache, ifNeedsReloading]);
 
+  /**
+   * Reloads the location data.
+   */
   const reloadData = async () => {
     setIfNeedsReloading(false);
     setCurrentMapCache({
       ...currentMapCache,
       loadedCoordinates: currentMapCache.selectedCoordinates,
     });
-    const responseData = await loadLocationData();
-    setData(responseData);
+    const responseData = await fetchLocationData();
+    if (responseData) {
+      setData(responseData);
+    }
   };
 
   return (
