+++ frontend/src/types/LocationDataTypes.tsx
@@ -1,0 +1,11 @@
+export interface LocationDataResponse {
+  currentDatasetData: DatasetItem[];
+  generalData: DatasetItem[];
+  extraRows: DatasetItem[];
+}
+
+export interface DatasetItem {
+  key: string;
+  value: string;
+  mapId: string;
+}
